+++ summoner-tui/src/Summoner/Tui/Widget.hs
@@ -17,11 +17,12 @@ module Summoner.Tui.Widget
        , listInBorder
        ) where
 
-import Brick.Types (Padding (Pad), Widget)
+import Brick (attrName)
+import Brick.Types (Widget)
 import Brick.Widgets.Border (borderWithLabel)
 import Brick.Widgets.Center (center)
-import Brick.Widgets.Core (hBox, hLimit, padLeftRight, padRight, str, txtWrap, vBox, vLimit,
-                           withAttr, (<+>))
+import Brick.Widgets.Core (Padding (Pad), hBox, hLimit, padLeftRight, padRight, str, txtWrap, vBox,
+                           vLimit, withAttr, (<+>))
 
 
 -- | Adds label to the Form's field.
@@ -91,4 +92,4 @@ listInBorder name limitH extraLimitV list = center
     $ borderLabel name
     $ center
     $ vBox
-    $ map (withAttr "blue-fg" . txtWrap . ("➤ " <>)) list
+    $ map (withAttr (attrName "blue-fg") . txtWrap . ("➤ " <>)) list